refactor: . Move remaining UI-related Quick Search code to QuickSearchTasksModal.ts
